requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.4.5
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
mysqlclient==2.1.1
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14